disatidjangoapi/requirements.txt
asgiref==3.8.1
click==8.1.7
Django==5.0.4
gunicorn==22.0.0
h11==0.14.0
packaging==24.0
psycopg2==2.9.9
sqlparse==0.5.0
typing_extensions==4.11.0
requests
django-cors-headers
djangorestframework
